+++ src/components/AuthorCard.astro
@@ -37,7 +37,7 @@ const socialLinks: SocialLink[] = [
 >
   <div class="flex flex-wrap gap-4">
     <Link
-      href={`/authors/${author.slug}`}
+      href={`/authors/${author.id}`}
       class={cn('block', linkDisabled && 'pointer-events-none')}
     >
       <AvatarComponent
